# public/openapi.pt-BR.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZAP CRM BR — API Pública",
    "version": "1.0.0",
    "summary": "API REST para operar uma instância auto-hospedada do ZAP CRM BR a partir dos seus próprios scripts e automações.",
...
    "contact": {
      "name": "ZAP CRM BR",
      "url": "https://github.com/brunopelatieri/zap-crm-br"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://{domain}/api/v1",
      "description": "Sua instância auto-hospedada do ZAP CRM BR",
      "variables": {
        "domain": {
          "default": "your-crm.example.com",
          "description": "O domínio onde sua instância do ZAP CRM BR está implantada."
        }
      }
    }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/messages": {
      "post": {
        "tags": ["Messages"],
        "summary": "Enviar uma mensagem de WhatsApp",
        "description": "Envia uma mensagem de WhatsApp para um número de telefone. Você informa um **número E.164**, não um id interno de conversa — o endpoint encontra-ou-cria o contato + conversa, e então envia.\n\n**`cold_send_limit` (429).** Num canal WhatsApp QRCode (Evolution) sem janela de sessão ativa de 24h da Meta, o envio para um contato frio é limitado por um teto diário/por hora por instância e um intervalo mínimo entre envios frios (SPEC 049 §6.2). Este caminho **sempre bloqueia** quando o limite é atingido — mais rígido que o inbox do painel, que só avisa um agente humano. Tente novamente após o header `Retry-After`.",
        "operationId": "sendMessage",
        "security": [{ "bearerAuth": ["messages:send"] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SendMessageRequest" },
              "examples": {
                "text": {
                  "summary": "Texto simples",
                  "value": {
                    "to": "+14155550123",
                    "type": "text",
                    "text": "Hi 👋"
                  }
                },
                "template": {
                  "summary": "Template aprovado",
                  "value": {
                    "to": "+14155550123",
                    "type": "template",
                    "template": {
                      "name": "order_update",
                      "language": "en_US",
                      "params": ["A123"]
                    }
                  }
                },
                "media": {
                  "summary": "Documento com legenda",
                  "value": {
                    "to": "+14155550123",
                    "type": "document",
                    "media_url": "https://example.com/invoice.pdf",
                    "filename": "invoice.pdf",
                    "text": "Here's your invoice"
                  }
                },
                "interactiveButtons": {
                  "summary": "Interativa — botões de resposta",
                  "value": {
                    "to": "+14155550123",
                    "type": "interactive",
                    "interactive_payload": {
                      "kind": "buttons",
                      "body": "How can we help?",
                      "buttons": [
                        { "id": "track_order", "title": "Track order" },
                        { "id": "talk_to_human", "title": "Talk to a human" }
                      ]
                    }
                  }
                },
                "interactiveList": {
                  "summary": "Interativa — lista",
                  "value": {
                    "to": "+14155550123",
                    "type": "interactive",
                    "interactive_payload": {
                      "kind": "list",
                      "body": "Pick a topic",
                      "button_label": "Choose",
                      "sections": [
                        {
                          "title": "Billing",
                          "rows": [
                            {
                              "id": "invoice",
                              "title": "Get an invoice",
                              "description": "PDF, sent by email"
                            }
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mensagem enviada.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/SendMessageResult" }
                  }
                },
                "example": {
                  "data": {
                    "message_id": "8f3c1e2a-0000-4000-8000-000000000010",
                    "whatsapp_message_id": "wamid.HBgLMTQxNTU1NTAxMjM=",
                    "conversation_id": "8f3c1e2a-0000-4000-8000-000000000011",
                    "contact_id": "8f3c1e2a-0000-4000-8000-000000000012",
                    "contact_created": true
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": {
            "description": "Limite de taxa por chave excedido (`rate_limited`), ou o limite de envio frio do WhatsApp QRCode foi atingido (`cold_send_limit`). Ambos trazem `Retry-After`.",
            "headers": {
              "Retry-After": { "$ref": "#/components/headers/RetryAfter" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "rateLimited": {
                    "value": {
                      "error": {
                        "code": "rate_limited",
                        "message": "Rate limit exceeded for this API key"
                      }
                    }
                  },
                  "coldSendLimit": {
                    "value": {
                      "error": {
                        "code": "cold_send_limit",
                        "message": "cold-send hourly limit reached (5/h)"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" },
          "502": {
            "description": "A requisição chegou à Meta e a Meta recusou o envio (`meta_error`).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "error": {
                    "code": "meta_error",
                    "message": "(#131026) Message undeliverable"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "wacrm_live_...",
        "description": "Uma chave de API criada no painel em Configurações → Chaves de API, enviada como `Authorization: Bearer wacrm_live_...`. Os escopos são aplicados por endpoint; veja o escopo exigido de cada operação."
      }
    },
...
    "headers": {
      "RetryAfter": {
        "description": "Segundos até o chamador poder tentar de novo.",
        "schema": { "type": "integer" }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Entrada malformada (`bad_request`).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": { "code": "bad_request", "message": "'to' is required" }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Chave de API ausente, malformada, desconhecida, revogada ou expirada (`unauthorized`). Estes casos são deliberadamente indistinguíveis no retorno.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": {
                "code": "unauthorized",
                "message": "Missing or invalid API key"
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "A chave é válida mas não tem o escopo que este endpoint exige (`forbidden`).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": {
                "code": "forbidden",
                "message": "This API key is missing the 'messages:send' scope"
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Erro de servidor (`internal`). A mensagem nunca vaza detalhe interno.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": {
                "code": "internal",
                "message": "Internal server error"
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string",
            "description": "Código de erro estável e comparável por máquina — trate a lógica sobre este campo, não sobre `message`."
          },
          "message": {
            "type": "string",
            "description": "Texto voltado ao humano, em inglês (ver Idioma das mensagens de erro). Pode ser reformulado; não faça correspondência de padrão sobre ele."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "$ref": "#/components/schemas/Error" } }
      },
...
      "InteractiveButton": {
        "type": "object",
        "required": ["id", "title"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Id estável, ecoado de volta quando o destinatário toca este botão."
          },
          "title": { "type": "string", "maxLength": 20 }
        }
      },
      "InteractiveButtonsPayload": {
        "type": "object",
        "required": ["kind", "body", "buttons"],
        "properties": {
          "kind": { "type": "string", "enum": ["buttons"] },
          "body": { "type": "string", "maxLength": 1024 },
          "header": { "type": "string", "maxLength": 60 },
          "footer": { "type": "string", "maxLength": 60 },
          "buttons": {
            "type": "array",
            "minItems": 1,
            "maxItems": 3,
            "items": { "$ref": "#/components/schemas/InteractiveButton" },
            "description": "1 a 3 botões. Os ids precisam ser únicos dentro da mensagem."
          }
        }
      },
      "InteractiveListRow": {
        "type": "object",
        "required": ["id", "title"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Id estável, ecoado de volta quando o destinatário seleciona esta linha."
          },
          "title": { "type": "string", "maxLength": 24 },
          "description": { "type": "string", "maxLength": 72 }
        }
      },
      "InteractiveListSection": {
        "type": "object",
        "required": ["rows"],
        "properties": {
          "title": { "type": "string" },
          "rows": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/InteractiveListRow" }
          }
        }
      },
      "InteractiveListPayload": {
        "type": "object",
        "required": ["kind", "body", "button_label", "sections"],
        "properties": {
          "kind": { "type": "string", "enum": ["list"] },
          "body": { "type": "string", "maxLength": 1024 },
          "header": { "type": "string", "maxLength": 60 },
          "footer": { "type": "string", "maxLength": 60 },
          "button_label": {
            "type": "string",
            "maxLength": 20,
            "description": "Rótulo do botão de tocar-para-expandir."
          },
          "sections": {
            "type": "array",
            "minItems": 1,
            "maxItems": 10,
            "items": { "$ref": "#/components/schemas/InteractiveListSection" },
            "description": "Até 10 seções; 1 a 10 linhas NO TOTAL entre todas as seções. Ids de linha precisam ser únicos em toda a mensagem."
          }
        }
      },
      "InteractiveMessagePayload": {
        "oneOf": [
          { "$ref": "#/components/schemas/InteractiveButtonsPayload" },
          { "$ref": "#/components/schemas/InteractiveListPayload" }
        ],
        "discriminator": {
          "propertyName": "kind",
          "mapping": {
            "buttons": "#/components/schemas/InteractiveButtonsPayload",
            "list": "#/components/schemas/InteractiveListPayload"
          }
        }
      },
      "TemplateSendParams": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Um nome de template aprovado pela Meta."
          },
          "language": {
            "type": "string",
            "default": "en_US",
            "example": "en_US"
          },
          "params": {
            "description": "Variáveis posicionais do corpo como um array, ou um objeto estruturado `{ header, body, buttons }`.",
            "oneOf": [
              { "type": "array", "items": { "type": "string" } },
              { "type": "object" }
            ]
          }
        }
      },
      "SendMessageCommon": {
        "type": "object",
        "required": ["to"],
        "properties": {
          "to": {
            "type": "string",
            "description": "Número de telefone em E.164. Encontra-ou-cria o contato + conversa.",
            "example": "+14155550123"
          },
          "reply_to_message_id": {
            "type": "string",
            "format": "uuid",
            "description": "Precisa ser uma mensagem na mesma conversa."
          },
          "channel_id": {
            "type": "string",
            "format": "uuid",
            "description": "De qual canal da conta enviar (SPEC 049 §5.4). Omita para usar o canal padrão da conta."
          },
          "name": {
            "type": "string",
            "description": "Nome a dar a um contato recém-criado. Ignorado se o contato já existir."
          }
        }
      },
      "TextMessageRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/SendMessageCommon" },
          {
            "type": "object",
            "required": ["text"],
            "properties": {
              "type": { "type": "string", "enum": ["text"], "default": "text" },
              "text": { "type": "string" }
            }
          }
        ]
      },
      "TemplateMessageRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/SendMessageCommon" },
          {
            "type": "object",
            "required": ["type", "template"],
            "properties": {
              "type": { "type": "string", "enum": ["template"] },
              "template": { "$ref": "#/components/schemas/TemplateSendParams" }
            }
          }
        ]
      },
      "MediaMessageRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/SendMessageCommon" },
          {
            "type": "object",
            "required": ["type", "media_url"],
            "properties": {
              "type": {
                "type": "string",
                "enum": ["image", "video", "document", "audio"]
              },
              "media_url": { "type": "string", "format": "uri" },
              "filename": {
                "type": "string",
                "description": "Nome do arquivo do documento."
              },
              "text": {
                "type": "string",
                "description": "Também funciona como a legenda da mídia."
              }
            }
          }
        ]
      },
      "InteractiveMessageRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/SendMessageCommon" },
          {
            "type": "object",
            "required": ["type", "interactive_payload"],
            "properties": {
              "type": { "type": "string", "enum": ["interactive"] },
              "interactive_payload": {
                "$ref": "#/components/schemas/InteractiveMessagePayload"
              }
            }
          }
        ]
      },
      "SendMessageRequest": {
        "oneOf": [
          { "$ref": "#/components/schemas/TextMessageRequest" },
          { "$ref": "#/components/schemas/TemplateMessageRequest" },
          { "$ref": "#/components/schemas/MediaMessageRequest" },
          { "$ref": "#/components/schemas/InteractiveMessageRequest" }
        ]
      },
      "SendMessageResult": {
        "type": "object",
        "required": [
          "message_id",
          "whatsapp_message_id",
          "conversation_id",
          "contact_id",
          "contact_created"
        ],
        "properties": {
          "message_id": { "type": "string", "format": "uuid" },
          "whatsapp_message_id": {
            "type": ["string", "null"],
            "description": "O id `wamid.*` da Meta."
          },
          "conversation_id": { "type": "string", "format": "uuid" },
          "contact_id": { "type": "string", "format": "uuid" },
          "contact_created": {
            "type": "boolean",
            "description": "True se `to` não correspondia a um contato existente."
          }
        }
      },
...
    }
  }
}