# src/app/api/whatsapp/send/route.ts
import { NextResponse } from 'next/server';
import { createClient } from '@/lib/supabase/server';
import {
  checkRateLimit,
  rateLimitResponse,
  RATE_LIMITS,
} from '@/lib/rate-limit';
import {
  sendMessageToConversation,
  validateSendMessageParams,
  SendMessageError,
} from '@/lib/whatsapp/send-message';
import { claimConversation } from '@/lib/inbox/assignment';
import { findOrCreateInboxConversation } from '@/lib/inbox/find-or-create-conversation';
import { resolveDefaultChannelId } from '@/lib/channels/send';

// The dashboard's outbound-send endpoint. It owns auth, per-user rate
// limiting, and the two ways the UI targets a thread — an existing
// `conversation_id` (inbox) or a `contact_id` (Contact detail →
// find-or-create the conversation). The actual Meta plumbing (validate
// → send → persist → pause flows) lives in the shared
// `sendMessageToConversation` core, which the public `/api/v1/messages`
// endpoint reuses. This route is a thin adapter: resolve the
// conversation, delegate, then map `SendMessageError` back onto the
// dashboard's internal `{ error }` shape.
export async function POST(request: Request) {
  try {
    const supabase = await createClient();

    const {
      data: { user },
      error: authError,
    } = await supabase.auth.getUser();

    if (authError || !user) {
      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
    }

    // Per-user rate limit. Bucket key is scoped to this route so
    // `/broadcast` has an independent budget.
    const limit = checkRateLimit(`send:${user.id}`, RATE_LIMITS.send);
    if (!limit.success) {
      return rateLimitResponse(limit);
    }

    // Resolve the caller's account_id. Every downstream lookup
    // (conversation, whatsapp_config, message_templates) is account-
    // scoped post-multi-user, so the previous `user_id` filters
    // returned nothing for teammates who didn't author the row.
    const { data: profile } = await supabase
      .from('profiles')
      .select('account_id')
      .eq('user_id', user.id)
      .maybeSingle();
    const accountId = profile?.account_id as string | undefined;
    if (!accountId) {
      return NextResponse.json(
        { error: 'Your profile is not linked to an account.' },
        { status: 403 }
      );
    }

    const body = await request.json();
    const {
      // `conversation_id` targets an existing thread (inbox). `contact_id`
      // lets a caller initiate from a contact that may have no conversation
      // yet (Contact detail → Send template) — we find-or-create one below.
      conversation_id: conversationIdInput,
      contact_id,
      // Canal explícito para o caminho `contact_id` (SPEC 049 §4.6) — o
      // seletor de canal informa isto quando a conta tem mais de um
      // canal apto a enviar o que está sendo mandado. Ausente = padrão
      // implícito de sempre (`resolveDefaultChannelId`), comportamento
      // idêntico ao de antes desta SPEC. Ignorado no caminho
      // `conversation_id` — o canal de uma thread existente é imutável
      // (trigger da 059), não algo que o chamador escolhe aqui.
      channel_id: explicitChannelId,
      message_type,
      content_text,
      media_url,
      // Caminho do objeto no bucket, mandado pelo composer junto da URL
      // (migração 040). É por ele que o envio assina a URL para a Meta e
      // que a bolha exibe a mídia depois — a URL pública deixou de ser
      // buscável quando o bucket virou privado.
      media_path,
      filename,
      template_name,
      template_language,
      template_params,
      template_message_params,
      interactive_payload,
      reply_to_message_id,
    } = body;

    if ((!conversationIdInput && !contact_id) || !message_type) {
      return NextResponse.json(
        {
          error:
            'Either conversation_id or contact_id, plus message_type, are required',
        },
        { status: 400 }
      );
    }

    // Validate the message shape up front — before the contact_id path
    // finds-or-creates a conversation — so an invalid payload 400s
    // without leaving an orphan empty conversation behind.
    try {
      validateSendMessageParams({
        messageType: message_type,
        contentText: content_text,
        mediaUrl: media_url,
        templateName: template_name,
        interactivePayload: interactive_payload,
      });
    } catch (err) {
      if (err instanceof SendMessageError) {
        return NextResponse.json(
          { error: err.message },
          { status: err.status }
        );
      }
      throw err;
    }

    // Resolve the target conversation. With `conversation_id` we load the
    // existing thread; with `contact_id` we find-or-create one for the
    // contact so a business-initiated template send (Contact detail view)
    // reuses the shared send core below.
    let conversationId: string | null = null;
    // Dono atual da thread, se houver. Decide se este envio REIVINDICA a
    // conversa (está na fila) ou apenas escreve nela (já é minha, ou sou
    // admin respondendo na de outro agente).
    let currentAssignee: string | null = null;

    if (conversationIdInput) {
      const { data, error: convError } = await supabase
        .from('conversations')
        .select('id, assigned_agent_id')
        .eq('id', conversationIdInput)
        .eq('account_id', accountId)
        .single();

      // Sob a RLS da 039, a conversa de outro agente é invisível — este
      // 404 é a resposta fail-closed correta, e genérica de propósito.
      if (convError || !data) {
        return NextResponse.json(
          { error: 'Conversation not found' },
          { status: 404 }
        );
      }
      conversationId = data.id;
      currentAssignee = data.assigned_agent_id ?? null;
    } else {
      // contact_id path: verify the contact is in this account first so a
      // caller can't open a conversation against someone else's contact.
      const { data: contactRow, error: contactErr } = await supabase
        .from('contacts')
        .select('id')
        .eq('id', contact_id)
        .eq('account_id', accountId)
        .maybeSingle();

      if (contactErr || !contactRow) {
        return NextResponse.json(
          { error: 'Contact not found' },
          { status: 404 }
        );
      }

      let channelId: string;
      if (explicitChannelId) {
        // Precisa pertencer a ESTA conta — senão um `channel_id` de
        // outra conta abriria conversa lá (mesma checagem que a 025 e
        // a RLS já fazem para `contact_id` acima, repetida aqui porque
        // `channels` não tem RLS sob o client anon que impeça o SELECT
        // por id sozinho).
        const { data: channelRow, error: channelErr } = await supabase
          .from('channels')
          .select('id')
          .eq('id', explicitChannelId)
          .eq('account_id', accountId)
          .maybeSingle();
        if (channelErr || !channelRow) {
          return NextResponse.json(
            { error: 'Channel not found' },
            { status: 400 }
          );
        }
        channelId = channelRow.id;
      } else {
        try {
          channelId = await resolveDefaultChannelId(supabase, accountId);
        } catch {
          return NextResponse.json(
            { error: 'WhatsApp not configured for this account.' },
            { status: 400 }
          );
        }
      }

      const resolved = await findOrCreateInboxConversation(
        supabase,
        accountId,
        user.id,
        contact_id,
        channelId
      );
      if (!resolved.ok) {
        return NextResponse.json(
          { error: resolved.error },
          { status: resolved.status }
        );
      }
      conversationId = resolved.id;
      currentAssignee = resolved.assignedAgentId;
    }

    if (!conversationId) {
      return NextResponse.json(
        { error: 'Conversation not found' },
        { status: 404 }
      );
    }

    // "Quem responde, assume."
    //
    // A ORDEM AQUI É IRREVERSÍVEL: reivindicar ANTES de falar com a Meta.
    // Invertida, o agente que perde a corrida já entregou a mensagem ao
    // cliente numa conversa que pertence a outro — e o WhatsApp não
    // desfaz entrega. Por isso o 409 aborta sem nunca chamar a Meta.
    //
    // Só reivindica quando a thread está na fila. Se já é minha, o claim
    // seria um no-op; se é de outro agente — situação visível apenas para
    // admin/owner — responder NÃO deve roubar a thread de quem atende.
    if (currentAssignee === null) {
      const claim = await claimConversation(supabase, conversationId);
      if (!claim.ok) {
        return NextResponse.json(
          { error: claim.error, code: claim.code },
          { status: claim.status }
        );
      }
    }

    // Delegate to the shared send core (validates, sends to Meta with
    // phone-variant retry, persists, pauses active flow runs). Its
    // `SendMessageError` carries a machine code + HTTP status; the
    // dashboard maps it to the internal `{ error }` shape.
    try {
      const result = await sendMessageToConversation(supabase, accountId, {
        conversationId,
        messageType: message_type,
        contentText: content_text,
        mediaUrl: media_url,
        mediaPath: media_path,
        filename,
        templateName: template_name,
        templateLanguage: template_language,
        templateParams: template_params,
        templateMessageParams: template_message_params,
        interactivePayload: interactive_payload,
        replyToMessageId: reply_to_message_id,
        // Autoria: este é o caminho com usuário logado. A API pública
        // (/api/v1/messages) omite e grava null — lá não há humano.
        senderId: user.id,
        // SPEC 049 §6.2, D-1: nunca bloqueia — só registra o consumo,
        // para a cota de envio frio descrever o número inteiro, não só
        // o motor.
        coldSendOrigin: 'human',
      });

      return NextResponse.json({
        success: true,
        message_id: result.messageId,
        whatsapp_message_id: result.whatsappMessageId,
      });
    } catch (err) {
      if (err instanceof SendMessageError) {
        return NextResponse.json(
          { error: err.message },
          { status: err.status }
        );
      }
      throw err;
    }
  } catch (error) {
    console.error('Error in WhatsApp send POST:', error);
    return NextResponse.json(
      { error: 'Failed to send message' },
      { status: 500 }
    );
  }
}

// `findOrCreateConversation` vivia aqui e foi extraída, sem mudança de
// comportamento, para `@/lib/inbox/find-or-create-conversation`
// (SPEC 056 F2): a rota de transferência de canal precisa das mesmas
// três decisões — cliente sob RLS, thread nascendo atribuída, e 409 em
// vez de re-resolver na colisão do índice único.